listener: Add separate thread for listener telemetry upload loop.
